+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 1.2.0
+
+Unreleased
+
+- Added auto suggestion from history in interactive mode.
+- Added `-i` option.
+- Removed the display of default profile in interactive mode.
+
 ## 1.1.0
 
 Released 2020-09-02